chore: Cache buildThreadContent
* Extract updateAgreeStatus to common function while at it

+++ app/src/main/java/com/huanchengfly/tieba/post/ui/widgets/compose/Search.kt
@@ -245,7 +245,7 @@ fun SearchThreadItem(
 
                 ThreadAgreeBtn(
                     hasAgree = false,
-                    agreeNum = StringUtil.tiebaNumToLong(item.likeNum).toInt(),
+                    agreeNum = StringUtil.tiebaNumToLong(item.likeNum),
                     modifier = Modifier.weight(1f)
                 )
